Major refactor to include new features and better API.
API is now DSL-style instead of using an options hash. This makes it much
easier to configure the middleware and allows new functionality to be
introduced more easily.

Simplified code by proxying more to the mail gem instead of having options
that then configure the mail gem.

Support for auto-responder when form processed. Support for specifying order
of fields if desired.